# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Live Chat Api",
    "description": "Live Chat api docs",
    "contact": {},
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:3008/",
      "description": "local"
    }
  ],
  "paths": {
    "/users/create": {
      "post": {
        "tags": [
          "User Management"
        ],
        "summary": "Create User",
        "description": "Creates User",
        "operationId": "CreateUser",
        "parameters": [
          {
            "name": "Content-Type",
            "in": "header",
            "description": "",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string",
              "enum": [
                "application/json"
              ]
            }
          }
        ],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "$ref": ""
              },
              "example": {
                "firstName": "Amit",
                "lastName": "Mahato",
                "type": "consumer"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": " Get the response based on scenarios appears below example schema",
            "content": {
              "application/json": {
                "example": {
                  "success": true,
                  "user": {
                    "firstName": "Amit",
                    "lastName": "Mahato",
                    "type": "consumer",
                    "_id": "297c6690-e4fb-46b8-84d9-48f2ad90a184",
                    "createdAt": "2022-03-29T03:12:44.887Z",
                    "updatedAt": "2022-03-29T03:12:44.887Z",
                    "__v": 0
                  }
                },
                "schema": {
                  "$ref": "#"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request errors",
            "content": {
              "application/json": {
                "example": [
                  {
                    "success": false,
                    "message": "Kindly fix the error(s)",
                    "errors": {
                      "type": "Possible value are consumer,support"
                    }
                  },
                  {
                    "success": false,
                    "message": "Kindly fix the error(s)",
                    "errors": {
                      "type": "please provide a string"
                    }
                  }
                ]
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "example": {
                  "success": false,
                  "error": {
                    "error": "No user with this id found"
                  }
                }
              }
            }
          }
        },
        "deprecated": false
      }
    },
...
  }
}